Allow overriding required ID rules with `(required) = false`
